app, cmd/gogio: add android foreground permission and service
This adds the permission android.permission.FOREGROUND_SERVICE and adds
GioForegroundService which creates the tray Notification necessary to
implement the Foreground Service. The package foreground includes the methods
Start and Stop, which on android, notifies the system that the program
will perform background work and that it shouldn't be killed. The
foreground service is stopped using the Stop method.

See https://developer.android.com/guide/components/foreground-services
and https://developer.android.com/training/notify-user/build-notification

Signed-off-by: Masala <masala@riseup.net>